Implements jestMatterRequest
