+++ packages/core/src/jestutils/jestHelpers.ts
@@ -71,8 +71,6 @@ import { Environment, Lifecycle, LogFormat as MatterLogFormat, LogLevel as Matte
 import { Endpoint, ServerNode, ServerNodeStore } from '@matter/node';
 import { AggregatorEndpoint } from '@matter/node/endpoints';
 import { MdnsService } from '@matter/protocol';
-import { ColorControl } from '@matter/types/clusters/color-control';
-import { LevelControl } from '@matter/types/clusters/level-control';
 import { DeviceTypeId, VendorId } from '@matter/types/datatype';
 // @matterbridge
 import { MATTER_STORAGE_DIR, NODE_STORAGE_DIR } from '@matterbridge/types';
@@ -1141,168 +1139,4 @@ export async function deleteDevice(owner: ServerNode<ServerNode.RootEndpoint> | 
   expect(device.construction.status).toBe(Lifecycle.Status.Destroyed);
   await flushAllEndpointNumberPersistence(targetServer, 3, pause);
   return true;
-}
-
-/**
- * Build a Matter LevelControl `MoveToLevelRequest` payload.
- *
- * Helper used in tests to create a correctly shaped request object, including
- * `optionsMask` and `optionsOverride`.
- *
- * @param {number} level Target level (Matter `uint8`; commonly 0–254).
- * @param {number} transitionTime Transition time (Matter `uint16`; commonly in 1/10s units).
- * @param {boolean} executeIfOff Whether the command should be executed even when the device is off.
- * @returns {LevelControl.MoveToLevelRequest} The request payload.
- */
-export function getMoveToLevelRequest(level: number, transitionTime: number, executeIfOff: boolean): LevelControl.MoveToLevelRequest {
-  const request: LevelControl.MoveToLevelRequest = {
-    level,
-    transitionTime,
-    optionsMask: { executeIfOff, coupleColorTempToLevel: false },
-    optionsOverride: { executeIfOff, coupleColorTempToLevel: false },
-  };
-  return request;
-}
-
-/**
- * Build a Matter ColorControl `MoveToColorTemperatureRequest` payload.
- *
- * @param {number} colorTemperatureMireds Target color temperature in mireds (micro reciprocal degrees).
- * @param {number} transitionTime Transition time (Matter `uint16`; commonly in 1/10s units).
- * @param {boolean} executeIfOff Whether the command should be executed even when the device is off.
- * @returns {ColorControl.MoveToColorTemperatureRequest} The request payload.
- */
-export function getMoveToColorTemperatureRequest(colorTemperatureMireds: number, transitionTime: number, executeIfOff: boolean): ColorControl.MoveToColorTemperatureRequest {
-  const request: ColorControl.MoveToColorTemperatureRequest = {
-    colorTemperatureMireds,
-    transitionTime,
-    optionsMask: { executeIfOff },
-    optionsOverride: { executeIfOff },
-  };
-  return request;
-}
-
-/**
- * Build a Matter ColorControl `MoveToHueRequest` payload.
- *
- * Uses `ColorControl.Direction.Shortest` by default.
- *
- * @param {number} hue Target hue (Matter `uint8`; commonly 0–254).
- * @param {number} transitionTime Transition time (Matter `uint16`; commonly in 1/10s units).
- * @param {boolean} executeIfOff Whether the command should be executed even when the device is off.
- * @returns {ColorControl.MoveToHueRequest} The request payload.
- */
-export function getMoveToHueRequest(hue: number, transitionTime: number, executeIfOff: boolean): ColorControl.MoveToHueRequest {
-  const request: ColorControl.MoveToHueRequest = {
-    hue,
-    transitionTime,
-    direction: ColorControl.Direction.Shortest,
-    optionsMask: { executeIfOff },
-    optionsOverride: { executeIfOff },
-  };
-  return request;
-}
-
-/**
- * Build a Matter ColorControl `EnhancedMoveToHueRequest` payload.
- *
- * Uses `ColorControl.Direction.Shortest` by default.
- *
- * @param {number} enhancedHue Target enhanced hue (Matter `uint16`; commonly 0–65535).
- * @param {number} transitionTime Transition time (Matter `uint16`; commonly in 1/10s units).
- * @param {boolean} executeIfOff Whether the command should be executed even when the device is off.
- * @returns {ColorControl.EnhancedMoveToHueRequest} The request payload.
- */
-export function getEnhancedMoveToHueRequest(enhancedHue: number, transitionTime: number, executeIfOff: boolean): ColorControl.EnhancedMoveToHueRequest {
-  const request: ColorControl.EnhancedMoveToHueRequest = {
-    enhancedHue,
-    transitionTime,
-    direction: ColorControl.Direction.Shortest,
-    optionsMask: { executeIfOff },
-    optionsOverride: { executeIfOff },
-  };
-  return request;
-}
-
-/**
- * Build a Matter ColorControl `MoveToSaturationRequest` payload.
- *
- * @param {number} saturation Target saturation (Matter `uint8`; commonly 0–254).
- * @param {number} transitionTime Transition time (Matter `uint16`; commonly in 1/10s units).
- * @param {boolean} executeIfOff Whether the command should be executed even when the device is off.
- * @returns {ColorControl.MoveToSaturationRequest} The request payload.
- */
-export function getMoveToSaturationRequest(saturation: number, transitionTime: number, executeIfOff: boolean): ColorControl.MoveToSaturationRequest {
-  const request: ColorControl.MoveToSaturationRequest = {
-    saturation,
-    transitionTime,
-    optionsMask: { executeIfOff },
-    optionsOverride: { executeIfOff },
-  };
-  return request;
-}
-
-/**
- * Build a Matter ColorControl `MoveToHueAndSaturationRequest` payload.
- *
- * @param {number} hue Target hue (Matter `uint8`; commonly 0–254).
- * @param {number} saturation Target saturation (Matter `uint8`; commonly 0–254).
- * @param {number} transitionTime Transition time (Matter `uint16`; commonly in 1/10s units).
- * @param {boolean} executeIfOff Whether the command should be executed even when the device is off.
- * @returns {ColorControl.MoveToHueAndSaturationRequest} The request payload.
- */
-export function getMoveToHueAndSaturationRequest(hue: number, saturation: number, transitionTime: number, executeIfOff: boolean): ColorControl.MoveToHueAndSaturationRequest {
-  const request: ColorControl.MoveToHueAndSaturationRequest = {
-    hue,
-    saturation,
-    transitionTime,
-    optionsMask: { executeIfOff },
-    optionsOverride: { executeIfOff },
-  };
-  return request;
-}
-
-/**
- * Build a Matter ColorControl `EnhancedMoveToHueAndSaturationRequest` payload.
- *
- * @param {number} enhancedHue Target enhanced hue (Matter `uint16`; commonly 0–65535).
- * @param {number} saturation Target saturation (Matter `uint8`; commonly 0–254).
- * @param {number} transitionTime Transition time (Matter `uint16`; commonly in 1/10s units).
- * @param {boolean} executeIfOff Whether the command should be executed even when the device is off.
- * @returns {ColorControl.EnhancedMoveToHueAndSaturationRequest} The request payload.
- */
-export function getEnhancedMoveToHueAndSaturationRequest(
-  enhancedHue: number,
-  saturation: number,
-  transitionTime: number,
-  executeIfOff: boolean,
-): ColorControl.EnhancedMoveToHueAndSaturationRequest {
-  const request: ColorControl.EnhancedMoveToHueAndSaturationRequest = {
-    enhancedHue,
-    saturation,
-    transitionTime,
-    optionsMask: { executeIfOff },
-    optionsOverride: { executeIfOff },
-  };
-  return request;
-}
-
-/**
- * Build a Matter ColorControl `MoveToColorRequest` payload.
- *
- * @param {number} colorX Target X coordinate in CIE 1931 XY color space (Matter `uint16`).
- * @param {number} colorY Target Y coordinate in CIE 1931 XY color space (Matter `uint16`).
- * @param {number} transitionTime Transition time (Matter `uint16`; commonly in 1/10s units).
- * @param {boolean} executeIfOff Whether the command should be executed even when the device is off.
- * @returns {ColorControl.MoveToColorRequest} The request payload.
- */
-export function getMoveToColorRequest(colorX: number, colorY: number, transitionTime: number, executeIfOff: boolean): ColorControl.MoveToColorRequest {
-  const request: ColorControl.MoveToColorRequest = {
-    colorX,
-    colorY,
-    transitionTime,
-    optionsMask: { executeIfOff },
-    optionsOverride: { executeIfOff },
-  };
-  return request;
 }
